+++ build.clj
@@ -17,7 +17,8 @@
    [clojure.tools.build.api :as b]))
 
 (def lib-coord 'org.suskalo/coffi)
-(def version (format "0.6.%s" (b/git-count-revs nil)))
+;;(def version (format "0.6.%s" (b/git-count-revs nil)))
+(def version "1.0.450")
 
 (def resource-dirs ["resources/"])
 
@@ -49,11 +50,14 @@
   "Compiles java classes required for interop."
   [opts]
   (.mkdirs (io/file class-dir))
-  (b/process {:command-args ["javac" "--enable-preview"
-                             "src/java/coffi/ffi/Loader.java"
-                             "-d" class-dir
-                             "-target" "19"
-                             "-source" "19"]})
+  (let [compilation-result
+        (b/process {:command-args ["javac"
+                                   "src/java/coffi/ffi/Loader.java"
+                                   "-d" class-dir
+                                   "-target" "22"
+                                   "-source" "22"]})]
+    (when-not (zero? (:exit compilation-result))
+      (b/delete {:path class-dir})))
   opts)
 
 (defn- write-pom