+++ src/clj/coffi/mem.clj
@@ -1,5 +1,5 @@
 (ns coffi.mem
-  "Functions for managing native allocations, memory sessions, and (de)serialization.
+  "Functions for managing native allocations, memory arenas, and (de)serialization.
 
   For any new type to be implemented, three multimethods must be overriden, but
   which three depends on the native representation of the type.
@@ -13,21 +13,17 @@
   struct, or array, then [[c-layout]] must be overriden to return the native
   layout of the type, and [[serialize-into]] and [[deserialize-from]] should be
   overriden to allow marshaling values of the type into and out of memory
-  segments.
-
-  When writing code that manipulates a segment, it's best practice to
-  use [[with-acquired]] on the [[segment-session]] in order to ensure it won't be
-  released during its manipulation."
+  segments."
   (:require
    [clojure.set :as set]
    [clojure.spec.alpha :as s])
   (:import
    (java.lang.foreign
-    Addressable
-    MemoryAddress
+    AddressLayout
+    Arena
     MemoryLayout
     MemorySegment
-    MemorySession
+    MemorySegment$Scope
     SegmentAllocator
     ValueLayout
     ValueLayout$OfByte
@@ -36,126 +32,71 @@
     ValueLayout$OfLong
     ValueLayout$OfChar
     ValueLayout$OfFloat
-    ValueLayout$OfDouble
-    ValueLayout$OfAddress)
+    ValueLayout$OfDouble)
    (java.lang.ref Cleaner)
+   (java.util.function Consumer)
    (java.nio ByteOrder)))
 
 (set! *warn-on-reflection* true)
 
-(defn stack-session
-  "Constructs a new session for use only in this thread.
+(defn confined-arena
+  "Constructs a new arena for use only in this thread.
 
-  The memory allocated within this session is cheap to allocate, like a native
-  stack."
-  (^MemorySession []
-   (MemorySession/openConfined))
-  (^MemorySession [^Cleaner cleaner]
-   (MemorySession/openConfined cleaner)))
+  The memory allocated within this arena is cheap to allocate, like a native
+  stack.
 
-(defn ^:deprecated stack-scope
-  "Constructs a new scope for use only in this thread.
+  The memory allocated within this arena will be cleared once it is closed, so
+  it is usually a good idea to create it in a [[with-open]] clause."
+  (^Arena []
+   (Arena/ofConfined)))
 
-  The memory allocated within this scope is cheap to allocate, like a native
-  stack."
-  ^MemorySession []
-  (stack-session))
+(defn shared-arena
+  "Constructs a new shared memory arena.
 
-(defn shared-session
-  "Constructs a new shared memory session.
+  This arena can be shared across threads and memory allocated in it will only
+  be cleaned up once any thread accessing the arena closes it."
+  (^Arena []
+   (Arena/ofShared)))
 
-  This session can be shared across threads and memory allocated in it will only
-  be cleaned up once every thread accessing the session closes it."
-  (^MemorySession []
-   (MemorySession/openShared))
-  (^MemorySession [^Cleaner cleaner]
-   (MemorySession/openShared cleaner)))
+(defn auto-arena
+  "Constructs a new memory arena that is managed by the garbage collector.
 
-(defn ^:deprecated shared-scope
-  "Constructs a new shared scope.
-
-  This scope can be shared across threads and memory allocated in it will only
-  be cleaned up once every thread accessing the scope closes it."
-  ^MemorySession []
-  (shared-session))
-
-(defn connected-session
-  "Constructs a new memory session to reclaim all connected resources at once.
-
-  The session may be shared across threads, and all resources created with it
-  will be cleaned up at the same time, when all references have been collected.
-
-  This type of session cannot be closed, and therefore should not be created in
-  a [[with-open]] clause."
-  ^MemorySession []
-  (MemorySession/openImplicit))
-
-(defn ^:deprecated connected-scope
-  "Constructs a new scope to reclaim all connected resources at once.
-
-  The scope may be shared across threads, and all resources created with it will
+  The arena may be shared across threads, and all resources created with it will
   be cleaned up at the same time, when all references have been collected.
 
-  This type of scope cannot be closed, and therefore should not be created in
+  This type of arena cannot be closed, and therefore should not be created in
   a [[with-open]] clause."
-  ^MemorySession []
-  (connected-session))
+  ^Arena []
+  (Arena/ofAuto))
 
-(defn global-session
-  "Constructs the global session, which will never reclaim its resources.
+(defn global-arena
+  "Constructs the global arena, which will never reclaim its resources.
 
-  This session may be shared across threads, but is intended mainly in cases
-  where memory is allocated with [[alloc]] but is either never freed or whose
-  management is relinquished to a native library, such as when returned from a
-  callback."
-  ^MemorySession []
-  (MemorySession/global))
-
-(defn ^:deprecated global-scope
-  "Constructs the global scope, which will never reclaim its resources.
-
-  This scope may be shared across threads, but is intended mainly in cases where
+  This arena may be shared across threads, but is intended mainly in cases where
   memory is allocated with [[alloc]] but is either never freed or whose
   management is relinquished to a native library, such as when returned from a
   callback."
-  ^MemorySession []
-  (global-session))
+  ^Arena []
+  (Arena/global))
 
-(defn session-allocator
-  "Constructs a segment allocator from the given `session`.
-
-  This is primarily used when working with unwrapped downcall functions. When a
-  downcall function returns a non-primitive type, it must be provided with an
-  allocator."
-  ^SegmentAllocator [^MemorySession session]
-  (SegmentAllocator/newNativeArena session))
-
-(defn ^:deprecated scope-allocator
-  "Constructs a segment allocator from the given `scope`.
+(defn arena-allocator
+  "Constructs a [[SegmentAllocator]] from the given [[Arena]].
 
   This is primarily used when working with unwrapped downcall functions. When a
   downcall function returns a non-primitive type, it must be provided with an
   allocator."
-  ^SegmentAllocator [^MemorySession scope]
-  (session-allocator scope))
-
-(defn segment-session
-  "Gets the memory session used to construct the `segment`."
-  ^MemorySession [segment]
-  (.session ^MemorySegment segment))
-
-(defn ^:deprecated segment-scope
-  "Gets the scope used to construct the `segment`."
-  ^MemorySession [segment]
-  (segment-session segment))
+  ^SegmentAllocator [^Arena arena]
+  (reify SegmentAllocator
+    (^MemorySegment allocate [_this ^long byte-size ^long byte-alignment]
+      (.allocate arena ^long byte-size ^long byte-alignment))))
 
 (defn alloc
   "Allocates `size` bytes.
 
-  If a `session` is provided, the allocation will be reclaimed when it is closed."
-  (^MemorySegment [size] (alloc size (connected-session)))
-  (^MemorySegment [size session] (MemorySegment/allocateNative (long size) ^MemorySession session))
-  (^MemorySegment [size alignment session] (MemorySegment/allocateNative (long size) (long alignment) ^MemorySession session)))
+  If an `arena` is provided, the allocation will be reclaimed when it is closed."
+  (^MemorySegment [size] (alloc size (auto-arena)))
+  (^MemorySegment [size arena] (.allocate ^Arena arena (long size)))
+  (^MemorySegment [size alignment arena] (.allocate ^Arena arena (long size) (long alignment))))
 
 (defn alloc-with
   "Allocates `size` bytes using the `allocator`."
@@ -164,47 +105,24 @@
   (^MemorySegment [allocator size alignment]
    (.allocate ^SegmentAllocator allocator (long size) (long alignment))))
 
-(defmacro with-acquired
-  "Acquires one or more `sessions` until the `body` completes.
-
-  This is only necessary to do on shared sessions, however if you are operating
-  on an arbitrary passed session, it is best practice to wrap code that
-  interacts with it wrapped in this."
-  {:style/indent 1}
-  [sessions & body]
-  (if (seq sessions)
-    `(let [session# ~(first sessions)
-           res# (volatile! ::invalid-value)]
-       (.whileAlive
-        ^MemorySession session#
-        (^:once fn* []
-         (vreset! res#
-                  (with-acquired [~@(rest sessions)]
-                    ~@body))))
-       @res#)
-    `(do ~@body)))
-(s/fdef with-acquired
-  :args (s/cat :sessions any?
-               :body (s/* any?)))
-
 (defn address-of
   "Gets the address of a given segment.
 
   This value can be used as an argument to functions which take a pointer."
-  ^MemoryAddress [addressable]
-  (.address ^Addressable addressable))
+  ^long [addressable]
+  (.address ^MemorySegment addressable))
 
 (defn null?
   "Checks if a memory address is null."
   [addr]
-  (or (.equals (MemoryAddress/NULL) addr) (not addr)))
+  (or (.equals (MemorySegment/NULL) addr) (not addr)))
 
 (defn address?
   "Checks if an object is a memory address.
 
   `nil` is considered an address."
   [addr]
-  (or (nil? addr) (instance? MemoryAddress addr)))
+  (or (nil? addr) (instance? MemorySegment addr)))
 
 (defn slice
   "Get a slice over the `segment` with the given `offset`."
@@ -213,46 +131,47 @@
   (^MemorySegment [segment offset size]
    (.asSlice ^MemorySegment segment (long offset) (long size))))
 
-(defn slice-into
-  "Get a slice into the `segment` starting at the `address`."
-  (^MemorySegment [address segment]
-   (.asSlice ^MemorySegment segment ^MemoryAddress address))
-  (^MemorySegment [address segment size]
-   (.asSlice ^MemorySegment segment ^MemoryAddress address (long size))))
+(defn reinterpret
+  "Reinterprets the `segment` as having the passed `size`.
 
-(defn with-offset
-  "Get a new address `offset` from the old `address`."
-  ^MemoryAddress [address offset]
-  (.addOffset ^MemoryAddress address (long offset)))
-
-(defn add-close-action!
-  "Adds a 0-arity function to be run when the `session` closes."
-  [^MemorySession session ^Runnable action]
-  (.addCloseAction session action)
-  nil)
+  If `arena` is passed, the scope of the `segment` is associated with the arena,
+  as well as its access constraints. If `cleanup` is passed, it will be a
+  1-argument function of a fresh memory segment backed by the same memory as the
+  returned segment which should perform any required cleanup operations. It will
+  be called when the `arena` is closed."
+  (^MemorySegment [^MemorySegment segment size]
+   (.reinterpret segment (long size) (auto-arena) nil))
+  (^MemorySegment [^MemorySegment segment size ^Arena arena]
+   (.reinterpret segment (long size) arena nil))
+  (^MemorySegment [^MemorySegment segment size ^Arena arena cleanup]
+   (.reinterpret segment (long size) arena
+                 (reify Consumer
+                   (accept [_this segment]
+                     (cleanup segment))))))
 
 (defn as-segment
-  "Dereferences an `address` into a memory segment associated with the `session`."
-  (^MemorySegment [^MemoryAddress address size]
-   (MemorySegment/ofAddress address (long size) (connected-session)))
-  (^MemorySegment [^MemoryAddress address size session]
-   (MemorySegment/ofAddress address (long size) session)))
+  "Dereferences an `address` into a memory segment associated with the `arena` (default global)."
+  (^MemorySegment [^long address]
+   (MemorySegment/ofAddress address))
+  (^MemorySegment [^long address size]
+   (reinterpret (MemorySegment/ofAddress address) size))
+  (^MemorySegment [^long address size ^Arena arena]
+   (reinterpret (MemorySegment/ofAddress address) (long size) arena nil))
+  (^MemorySegment [^long address size ^Arena arena cleanup]
+   (reinterpret (MemorySegment/ofAddress address) (long size) arena cleanup)))
 
 (defn copy-segment
   "Copies the content to `dest` from `src`.
 
   Returns `dest`."
   ^MemorySegment [^MemorySegment dest ^MemorySegment src]
-  (with-acquired [(segment-session src) (segment-session dest)]
-    (.copyFrom dest src)
-    dest))
+  (.copyFrom dest src))
 
 (defn clone-segment
   "Clones the content of `segment` into a new segment of the same size."
-  (^MemorySegment [segment] (clone-segment segment (connected-session)))
-  (^MemorySegment [^MemorySegment segment session]
-   (with-acquired [(segment-session segment) session]
-     (copy-segment ^MemorySegment (alloc (.byteSize segment) session) segment))))
+  (^MemorySegment [segment] (clone-segment segment (auto-arena)))
+  (^MemorySegment [^MemorySegment segment ^Arena arena]
+   (copy-segment ^MemorySegment (alloc (.byteSize segment) arena) segment)))
 
 (defn slice-segments
   "Constructs a lazy seq of `size`-length memory segments, sliced from `segment`."
@@ -307,7 +226,7 @@
   "The [[MemoryLayout]] for a c-sized double in [[native-endian]] [[ByteOrder]]."
   ValueLayout/JAVA_DOUBLE)
 
-(def ^ValueLayout$OfAddress pointer-layout
+(def ^AddressLayout pointer-layout
   "The [[MemoryLayout]] for a native pointer in [[native-endian]] [[ByteOrder]]."
   ValueLayout/ADDRESS)
 
@@ -517,20 +436,20 @@
    (.get segment (.withOrder ^ValueLayout$OfDouble double-layout byte-order) offset)))
 
 (defn read-address
-  "Reads a [[MemoryAddress]] from the `segment`, at an optional `offset`."
+  "Reads an address from the `segment`, at an optional `offset`, wrapped in a [[MemorySegment]]."
   {:inline
    (fn read-address-inline
      ([segment]
       `(let [segment# ~segment]
-         (.get ^MemorySegment segment# ^ValueLayout$OfAddress pointer-layout 0)))
+         (.get ^MemorySegment segment# ^AddressLayout pointer-layout 0)))
      ([segment offset]
       `(let [segment# ~segment
              offset# ~offset]
-         (.get ^MemorySegment segment# ^ValueLayout$OfAddress pointer-layout offset#))))}
-  (^MemoryAddress [^MemorySegment segment]
-   (.get segment ^ValueLayout$OfAddress pointer-layout 0))
-  (^MemoryAddress [^MemorySegment segment ^long offset]
-   (.get segment ^ValueLayout$OfAddress pointer-layout offset)))
+         (.get ^MemorySegment segment# ^AddressLayout pointer-layout offset#))))}
+  (^MemorySegment [^MemorySegment segment]
+   (.get segment ^AddressLayout pointer-layout 0))
+  (^MemorySegment [^MemorySegment segment ^long offset]
+   (.get segment ^AddressLayout pointer-layout offset)))
 
 (defn write-byte
   "Writes a [[byte]] to the `segment`, at an optional `offset`."
@@ -715,22 +634,22 @@
    (.set segment (.withOrder ^ValueLayout$OfDouble double-layout byte-order) offset value)))
 
 (defn write-address
-  "Writes a [[MemoryAddress]] to the `segment`, at an optional `offset`."
+  "Writes the address of the [[MemorySegment]] `value` to the `segment`, at an optional `offset`."
   {:inline
    (fn write-address-inline
      ([segment value]
       `(let [segment# ~segment
              value# ~value]
-         (.set ^MemorySegment segment# ^ValueLayout$OfAddress pointer-layout 0 ^Addressable value#)))
+         (.set ^MemorySegment segment# ^AddressLayout pointer-layout 0 ^MemorySegment value#)))
      ([segment offset value]
       `(let [segment# ~segment
              offset# ~offset
              value# ~value]
-         (.set ^MemorySegment segment# ^ValueLayout$OfAddress pointer-layout offset# ^Addressable value#))))}
-  (^MemoryAddress [^MemorySegment segment ^MemoryAddress value]
-   (.set segment ^ValueLayout$OfAddress pointer-layout 0 value))
-  (^MemoryAddress [^MemorySegment segment ^long offset ^MemoryAddress value]
-   (.set segment ^ValueLayout$OfAddress pointer-layout offset value)))
+         (.set ^MemorySegment segment# ^AddressLayout pointer-layout offset# ^MemorySegment value#))))}
+  ([^MemorySegment segment ^MemorySegment value]
+   (.set segment ^AddressLayout pointer-layout 0 value))
+  ([^MemorySegment segment ^long offset ^MemorySegment value]
+   (.set segment ^AddressLayout pointer-layout offset value)))
 
 (defn- type-dispatch
   "Gets a type dispatch value from a (potentially composite) type."
@@ -867,7 +786,7 @@
    ::char Byte/TYPE
    ::float Float/TYPE
    ::double Double/TYPE
-   ::pointer MemoryAddress
+   ::pointer MemorySegment
    ::void Void/TYPE})
 
 (defn java-layout
@@ -894,8 +813,8 @@
 
 (defn alloc-instance
   "Allocates a memory segment for the given `type`."
-  (^MemorySegment [type] (alloc-instance type (connected-session)))
-  (^MemorySegment [type session] (MemorySegment/allocateNative ^long (size-of type) ^MemorySession session)))
+  (^MemorySegment [type] (alloc-instance type (auto-arena)))
+  (^MemorySegment [type arena] (.allocate ^Arena arena ^long (size-of type) ^long (align-of type))))
 
 (declare serialize serialize-into)
 
@@ -903,136 +822,130 @@
   "Constructs a serialized version of the `obj` and returns it.
 
   Any new allocations made during the serialization should be tied to the given
-  `session`, except in extenuating circumstances.
+  `arena`, except in extenuating circumstances.
 
   This method should only be implemented for types that serialize to primitives."
   (fn
     #_{:clj-kondo/ignore [:unused-binding]}
-    [obj type session]
+    [obj type arena]
     (type-dispatch type)))
 
 (defmethod serialize* :default
-  [obj type _session]
+  [obj type _arena]
   (throw (ex-info "Attempted to serialize a non-primitive type with primitive methods"
                   {:type type
                    :object obj})))
 
 (defmethod serialize* ::byte
-  [obj _type _session]
+  [obj _type _arena]
   (byte obj))
 
 (defmethod serialize* ::short
-  [obj _type _session]
+  [obj _type _arena]
   (short obj))
 
 (defmethod serialize* ::int
-  [obj _type _session]
+  [obj _type _arena]
   (int obj))
 
 (defmethod serialize* ::long
-  [obj _type _session]
+  [obj _type _arena]
   (long obj))
 
 (defmethod serialize* ::char
-  [obj _type _session]
+  [obj _type _arena]
   (char obj))
 
 (defmethod serialize* ::float
-  [obj _type _session]
+  [obj _type _arena]
   (float obj))
 
 (defmethod serialize* ::double
-  [obj _type _session]
+  [obj _type _arena]
   (double obj))
 
 (defmethod serialize* ::pointer
-  [obj type session]
+  [obj type arena]
   (if-not (null? obj)
     (if (sequential? type)
-      (with-acquired [session]
-        (let [segment (alloc-instance (second type) session)]
-          (serialize-into obj (second type) segment session)
-          (address-of segment)))
+      (let [segment (alloc-instance (second type) arena)]
+        (serialize-into obj (second type) segment arena)
+        (address-of segment))
       obj)
-    (MemoryAddress/NULL)))
+    (MemorySegment/NULL)))
 
 (defmethod serialize* ::void
-  [_obj _type _session]
+  [_obj _type _arena]
   nil)
 
 (defmulti serialize-into
   "Writes a serialized version of the `obj` to the given `segment`.
 
   Any new allocations made during the serialization should be tied to the given
-  `session`, except in extenuating circumstances.
+  `arena`, except in extenuating circumstances.
 
   This method should be implemented for any type which does not
   override [[c-layout]].
 
   For any other type, this will serialize it as [[serialize*]] before writing
-  the result value into the `segment`.
-
-  Implementations of this should be inside a [[with-acquired]] block for the
-  `session` if they perform multiple memory operations."
+  the result value into the `segment`."
   (fn
     #_{:clj-kondo/ignore [:unused-binding]}
-    [obj type segment session]
+    [obj type segment arena]
     (type-dispatch type)))
 
 (defmethod serialize-into :default
-  [obj type segment session]
+  [obj type segment arena]
   (if-some [prim-layout (primitive-type type)]
-    (with-acquired [(segment-session segment) session]
-      (serialize-into (serialize* obj type session) prim-layout segment session))
+    (serialize-into (serialize* obj type arena) prim-layout segment arena)
     (throw (ex-info "Attempted to serialize an object to a type that has not been overridden"
                     {:type type
                      :object obj}))))
 
 (defmethod serialize-into ::byte
-  [obj _type segment _session]
+  [obj _type segment _arena]
   (write-byte segment (byte obj)))
 
 (defmethod serialize-into ::short
-  [obj type segment _session]
+  [obj type segment _arena]
   (if (sequential? type)
     (write-short segment 0 (second type) (short obj))
     (write-short segment (short obj))))
 
 (defmethod serialize-into ::int
-  [obj type segment _session]
+  [obj type segment _arena]
   (if (sequential? type)
     (write-int segment 0 (second type) (int obj))
     (write-int segment (int obj))))
 
 (defmethod serialize-into ::long
-  [obj type segment _session]
+  [obj type segment _arena]
   (if (sequential? type)
     (write-long segment 0 (second type) (long obj))
     (write-long segment (long obj))))
 
 (defmethod serialize-into ::char
-  [obj _type segment _session]
+  [obj _type segment _arena]
   (write-char segment (char obj)))
 
 (defmethod serialize-into ::float
-  [obj type segment _session]
+  [obj type segment _arena]
   (if (sequential? type)
     (write-float segment 0 (second type) (float obj))
     (write-float segment (float obj))))
 
 (defmethod serialize-into ::double
-  [obj type segment _session]
+  [obj type segment _arena]
   (if (sequential? type)
     (write-double segment 0 (second type) (double obj))
     (write-double segment (double obj))))
 
 (defmethod serialize-into ::pointer
-  [obj type segment session]
-  (with-acquired [(segment-session segment) session]
-    (write-address
-     segment
-     (cond-> obj
-       (sequential? type) (serialize* type session)))))
+  [obj type segment arena]
+  (write-address
+   segment
+   (cond-> obj
+     (sequential? type) (serialize* type arena))))
 
 (defn serialize
   "Serializes an arbitrary type.
@@ -1040,12 +953,12 @@
   For types which have a primitive representation, this serializes into that
   representation. For types which do not, it allocates a new segment and
   serializes into that."
-  ([obj type] (serialize obj type (connected-session)))
-  ([obj type session]
+  ([obj type] (serialize obj type (auto-arena)))
+  ([obj type arena]
    (if (primitive-type type)
-     (serialize* obj type session)
-     (let [segment (alloc-instance type session)]
-       (serialize-into obj type segment session)
+     (serialize* obj type arena)
+     (let [segment (alloc-instance type arena)]
+       (serialize-into obj type segment arena)
        segment))))
 
 (declare deserialize deserialize*)
@@ -1054,10 +967,7 @@
   "Deserializes the given segment into a Clojure data structure.
 
   For types that serialize to primitives, a default implementation will
-  deserialize the primitive before calling [[deserialize*]].
-
-  Implementations of this should be inside a [[with-acquired]] block for the the
-  `segment`'s session if they perform multiple memory operations."
+  deserialize the primitive before calling [[deserialize*]]."
   (fn
     #_{:clj-kondo/ignore [:unused-binding]}
     [segment type]
@@ -1113,9 +1023,8 @@
 
 (defmethod deserialize-from ::pointer
   [segment type]
-  (with-acquired [(segment-session segment)]
-    (cond-> (read-address segment)
-      (sequential? type) (deserialize* type))))
+  (cond-> (read-address segment)
+    (sequential? type) (deserialize* type)))
 
 (defmulti deserialize*
   "Deserializes a primitive object into a Clojure data structure.
@@ -1165,8 +1074,11 @@
   [addr type]
   (when-not (null? addr)
     (if (sequential? type)
-      (deserialize-from (as-segment addr (size-of (second type)))
-                        (second type))
+      (let [target-type (second type)]
+        (deserialize-from
+         (.reinterpret ^MemorySegment (read-address addr)
+                       ^long (size-of target-type))
+         target-type))
       addr)))
 
 (defmethod deserialize* ::void
@@ -1188,8 +1100,7 @@
 (defn seq-of
   "Constructs a lazy sequence of `type` elements deserialized from `segment`."
   [type segment]
-  (with-acquired [(segment-session segment)]
-    (map #(deserialize % type) (slice-segments segment (size-of type)))))
+  (map #(deserialize % type) (slice-segments segment (size-of type))))
 
 ;;; Raw composite types
 ;; TODO(Joshua): Ensure that all the raw values don't have anything happen on
@@ -1200,7 +1111,7 @@
   (c-layout type))
 
 (defmethod serialize-into ::raw
-  [obj _type segment _session]
+  [obj _type segment _arena]
   (if (instance? MemorySegment obj)
     (copy-segment segment obj)
     obj))
@@ -1218,15 +1129,15 @@
   ::pointer)
 
 (defmethod serialize* ::c-string
-  [obj _type session]
+  [obj _type ^Arena arena]
   (if obj
-    (address-of (.allocateUtf8String (session-allocator session) ^String obj))
-    (MemoryAddress/NULL)))
+    (.allocateFrom arena ^String obj)
+    (MemorySegment/NULL)))
 
 (defmethod deserialize* ::c-string
   [addr _type]
   (when-not (null? addr)
-    (.getUtf8String ^MemoryAddress addr 0)))
+    (.getString (.reinterpret ^MemorySegment addr Integer/MAX_VALUE) 0)))
 
 ;;; Union types
 
@@ -1237,7 +1148,7 @@
      (into-array MemoryLayout items))))
 
 (defmethod serialize-into ::union
-  [obj [_union _types & {:keys [dispatch extract]} :as type] segment session]
+  [obj [_union _types & {:keys [dispatch extract]} :as type] segment arena]
   (when-not dispatch
     (throw (ex-info "Attempted to serialize a union with no dispatch function"
                     {:type type
@@ -1249,7 +1160,7 @@
        obj)
      type
      segment
-     session)))
+     arena)))
 
 (defmethod deserialize-from ::union
   [segment type]
@@ -1266,7 +1177,7 @@
      (into-array MemoryLayout fields))))
 
 (defmethod serialize-into ::struct
-  [obj [_struct fields] segment session]
+  [obj [_struct fields] segment arena]
   (loop [offset 0
          fields fields]
     (when (seq fields)
@@ -1274,7 +1185,7 @@
             size (size-of type)]
         (serialize-into
          (get obj field) type
-         (slice segment offset size) session)
+         (slice segment offset size) arena)
         (recur (long (+ offset size)) (rest fields))))))
 
 (defmethod deserialize-from ::struct
@@ -1297,10 +1208,10 @@
 
 (defmethod c-layout ::padding
   [[_padding size]]
-  (MemoryLayout/paddingLayout (* 8 size)))
+  (MemoryLayout/paddingLayout size))
 
 (defmethod serialize-into ::padding
-  [_obj [_padding _size] _segment _session]
+  [_obj [_padding _size] _segment _arena]
   nil)
 
 (defmethod deserialize-from ::padding
@@ -1316,9 +1227,9 @@
    (c-layout type)))
 
 (defmethod serialize-into ::array
-  [obj [_array type count] segment session]
+  [obj [_array type count] segment arena]
   (dorun
-   (map #(serialize-into %1 type %2 session)
+   (map #(serialize-into %1 type %2 arena)
         obj
         (slice-segments (slice segment 0 (* count (size-of type)))
                         (size-of type)))))
@@ -1363,10 +1274,10 @@
       variants))))
 
 (defmethod serialize* ::enum
-  [obj [_enum variants & {:keys [repr]}] session]
+  [obj [_enum variants & {:keys [repr]}] arena]
   (serialize* ((enum-variants-map variants) obj)
               (or repr ::int)
-              session))
+              arena))
 
 (defmethod deserialize* ::enum
   [obj [_enum variants & {:keys [_repr]}]]
@@ -1381,9 +1292,9 @@
     ::int))
 
 (defmethod serialize* ::flagset
-  [obj [_flagset bits & {:keys [repr]}] session]
+  [obj [_flagset bits & {:keys [repr]}] arena]
   (let [bits-map (enum-variants-map bits)]
-    (reduce #(bit-set %1 (get bits-map %2)) (serialize* 0 (or repr ::int) session) obj)))
+    (reduce #(bit-set %1 (get bits-map %2)) (serialize* 0 (or repr ::int) arena) obj)))
 
 (defmethod deserialize* ::flagset
   [obj [_flagset bits & {:keys [repr]}]]
@@ -1415,8 +1326,8 @@
          [_type#]
          (primitive-type aliased#))
        (defmethod serialize* ~new-type
-         [obj# _type# session#]
-         (serialize* obj# aliased# session#))
+         [obj# _type# arena#]
+         (serialize* obj# aliased# arena#))
        (defmethod deserialize* ~new-type
          [obj# _type#]
          (deserialize* obj# aliased#)))
@@ -1425,8 +1336,8 @@
          [_type#]
          (c-layout aliased#))
        (defmethod serialize-into ~new-type
-         [obj# _type# segment# session#]
-         (serialize-into obj# aliased# segment# session#))
+         [obj# _type# segment# arena#]
+         (serialize-into obj# aliased# segment# arena#))
        (defmethod deserialize-from ~new-type
          [segment# _type#]
          (deserialize-from segment# aliased#)))))